refactor: move functions to same place they're used
